feat(hdmi-in): wire live-input mode into the locked MNIST HDMI demo
Day 2 deliverables — completes the SW side. Vivado-side work
(scripts/add_hdmi_in_pynqz2.tcl) is the only remaining blocker before
silicon bring-up.

07_sw/baremetal/demo_mnist_hdmi_bm.c
  - #ifdef LIVE_INPUT guard around all new code
  - At init: hdmi_in_init() after the HDMI-out hdmi_init()
  - Inside the main loop: when hdmi_in_locked() returns true, replace
    sweep_input28[i] with a live 28×28 captured + downsampled frame
  - Falls back to fixture if HDMI-RX isn't locked (no signal yet)
  - label = -1 in live mode (no ground truth); accuracy bookkeeping
    is gated behind `if (label >= 0)` so counters only advance in
    fixture mode
  - fbm_draw_image28 shows the LIVE input on the output panel when
    in live mode (so the user sees what the CGRA is consuming)

07_sw/baremetal/frame_to_mnist.{h,c}
  - Output type corrected from int16_t to uint8_t.
  - Reason: sweep_input28[] in mnist_sweep_fixture.h is
    `const uint8_t[100][784]` (0–255 range, MNIST polarity), and
    arm_cnn_vfp_run() takes uint8_t input. The earlier int16_t
    quantization with [-16384,16384] range was incompatible.
  - Pipeline reduces to: 28×28 box-filter → BGR-to-Y luma →
    invert (255 - Y). Result is plain 0–255 ready to feed.

07_sw/baremetal/Makefile
  - New build-time switch: make mnist_hdmi LIVE_INPUT=1
  - Conditionally appends hdmi_in_bm.c + frame_to_mnist.c to the
    MNIST_HDMI_CORE_SRCS list and adds -DLIVE_INPUT to the CFLAGS.

Build with: `cd 07_sw/baremetal && make BOARD=PYNQ_BASE LIVE_INPUT=1 mnist_hdmi`

Will fault on first MMIO read until the new bitstream
(scripts/add_hdmi_in_pynqz2.tcl) is built and loaded — but compiles
cleanly today, which validates the type-checking + API contract.

+++ 07_sw/baremetal/frame_to_mnist.h
@@ -25,24 +25,27 @@ typedef struct {
 #define HDMI_ROI_DEFAULT  ((hdmi_roi_t){.x = 180, .y = 100, .w = 280, .h = 280})
 
 /**
- * Downsample a region of the captured frame into a 28×28 INT16 tensor.
+ * Downsample a region of the captured frame into a 28×28 UINT8 tensor.
+ * The output format matches `sweep_input28[]` in mnist_sweep_fixture.h —
+ * a plain `uint8_t[784]` with 0 = MNIST-black background, 255 = white-ink
+ * digit — which `arm_cnn_vfp_run()` consumes directly.
  *
  * Pipeline:
  *   1. For each of 784 output pixels, compute the BGR average over the
  *      (w/28)×(h/28) source-pixel block.
  *   2. Convert to luma:  Y = (66*R + 129*G + 25*B + 128) >> 8.
- *   3. Invert:           Y = 255 - Y.   (Paint black-on-white → MNIST polarity)
- *   4. Centre on zero:   out = ((Y - 128) << 7).
- *      Matches the INT16 scaling that cnn_eval/emit_sweep_fixture.py applies
- *      to sweep_input28[]: signed [-32768..32767] range with ~128 as neutral.
+ *   3. Invert:           out = 255 - Y.
+ *      (Paint draws black ink on a white canvas; MNIST is white digit
+ *      on black background — invert in software so the existing CGRA
+ *      weights work without retraining.)
  *
  * @param fb      base address of the 640×480 BGR frame (from
  *                hdmi_in_current_frame()).
  * @param roi     region of interest within the frame.
- * @param out784  output buffer of 28*28 = 784 int16_t values.
+ * @param out784  output buffer of 28*28 = 784 uint8_t values.
  */
 void downsample_roi_to_mnist(const uint8_t *fb,
                               hdmi_roi_t      roi,
-                              int16_t         out784[28*28]);
+                              uint8_t         out784[28*28]);
 
 #endif /* FRAME_TO_MNIST_H */
